requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
multidict==6.0.4
psycopg2-binary==2.9.6
pydantic==1.10.7
pyee==8.2.2
pyppeteer==1.0.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
schedule==1.2.0
six==1.16.0
SQLAlchemy==2.0.13
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
websockets==10.4
yarl==1.9.2
zipp==3.15.0
